@@ -1,17 +1,19 @@
 name: supabase_flutter_example
-description: Demonstrates how to use supabase_flutter
+description: "Examples app showcasing usage of Supabase Flutter library"
 
 publish_to: 'none'
 
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.15.0 <3.0.0'
+  sdk: ^3.7.2
 
 dependencies:
   flutter:
     sdk: flutter
   supabase_flutter: ^2.9.1
+  cupertino_icons: ^1.0.8
+  flutter_facebook_auth: ^7.1.1
 
 dev_dependencies:
   flutter_test:
